requirements.txt
altgraph==0.17.4
charset-normalizer==3.4.3
colorama==0.4.6
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-barcode==0.16.1
pywin32-ctypes==0.2.3
qrcode==8.2
QtAwesome==1.4.0
QtPy==2.4.3
reportlab==4.4.3
